+++ packages/expo-router/src/react-navigation/core/usePreventBeforeUnload.tsx
@@ -1,0 +1,5 @@
+const NOOP = () => {};
+
+export function usePreventBeforeUnload(_preventBeforeUnload: boolean) {
+  return NOOP;
+}